requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-telegram-bot>=22.0,<23.0
SQLAlchemy>=2.0.36,<3.0.0
alembic>=1.16.0,<2.0.0
asyncpg>=0.30.0,<1.0.0
aiosqlite>=0.21.0,<1.0.0
redis>=5.0.0,<6.0.0
celery>=5.4.0,<6.0.0
python-dotenv>=1.1.0,<2.0.0
pydantic-settings>=2.7.0,<3.0.0
email-validator>=2.2.0,<3.0.0
APScheduler>=3.11.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
PyJWT>=2.10.0,<3.0.0
prometheus-client>=0.22.0,<1.0.0
matplotlib>=3.10.0,<4.0.0
jinja2>=3.1.6,<4.0.0
itsdangerous>=2.2.0,<3.0.0
httpx>=0.28.1,<1.0.0
orjson>=3.10.0,<4.0.0
tenacity>=9.0.0,<10.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
cryptography>=44.0.0,<45.0.0
